% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_schema.R
\name{create_schema}
\alias{create_schema}
\title{EXPLANATION OF FUNCTION}
\usage{
create_schema(tables = c("mtd", "ind_dat", "ind_boro_dat"))
}
\arguments{
\item{tables}{DETAILS ABOUT THE tables PARAMETER}
}
\value{
TRUE if it works
}
\description{
EXPLANATION OF FUNCTION
}
\examples{
EXAMPLE OF USAGE
import RPostgres
import RSQLite
import DBI
import xlsx
import yaml

}
