% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_log.R
\name{error_log}
\alias{error_log}
\title{Function to run on error when inserting data into the database}
\usage{
error_log(e, dtst)
}
\arguments{
\item{dtst}{The dataset from which the error occurred}
}
\value{
Logical - TRUE for worked ok.
}
\description{
Function to run on error when inserting data into the database
}
\examples{
dat <- data.frame()
insert_db(dat)
}
