% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_metadata.R
\name{get_metadata}
\alias{get_metadata}
\title{get all metadata}
\usage{
get_metadata()
}
\value{
All metadata in a data.frame
}
\description{
get all metadata
}
\examples{
get_metadata()

import RSQLite
}
