% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_cg_AQ.R
\name{insert_cg_AQ}
\alias{insert_cg_AQ}
\title{Insert Air Quality data into the database}
\usage{
insert_cg_AQ(log = "")
}
\arguments{
\item{log}{log file to save output to - defaults to}
}
\value{
Logical - TRUE for worked ok.
}
\description{
Insert Air Quality data into the database
}
\examples{
insert_cg_AQ

}
