% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_cg_datastore.R
\name{insert_cg_datastore}
\alias{insert_cg_datastore}
\title{Insert datastore pages from the cleaner, greener outcome}
\usage{
insert_cg_datastore(log = "")
}
\arguments{
\item{log}{log file to save output to - defaults to}
}
\value{
Logical - TRUE for worked ok.
}
\description{
Insert datastore pages from the cleaner, greener outcome
}
\examples{
insert_cg_AQ

}
