% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_cg_other.R
\name{insert_cg_other}
\alias{insert_cg_other}
\title{Inserts other cleaner greener data}
\usage{
insert_cg_other(log = "")
}
\arguments{
\item{log}{log file to save output to - defaults to}
}
\value{
Logical - TRUE for worked ok.
}
\description{
Inserts other cleaner greener data
}
\examples{
run_updates()


}
