% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_communities_datastore.R
\name{insert_communities_datastore}
\alias{insert_communities_datastore}
\title{Inserts communities datastore data}
\usage{
insert_communities_datastore(log = "")
}
\value{
Logical - TRUE for worked ok.
}
\description{
Inserts communities datastore data
}
\examples{
run_updates()

}
