% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_communities_socpol.R
\name{insert_communities_socpol}
\alias{insert_communities_socpol}
\title{Inserts other communities data (Crime - random health etc)}
\usage{
insert_communities_socpol(log = "")
}
\arguments{
\item{log}{log file to save output to - defaults to}
}
\value{
Logical - TRUE for worked ok.
}
\description{
Inserts other communities data (Crime - random health etc)
}
\examples{
run_updates()




}
