% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_db.R
\name{insert_db}
\alias{insert_db}
\title{Add data to database}
\usage{
insert_db(dat)
}
\arguments{
\item{dat}{A data.frame with correct columns and at least one row}
}
\value{
Logical - TRUE for worked ok.
}
\description{
Add data to database
}
\examples{
dat <- data.frame()
insert_db(dat)
}
