% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_econ_datastore.R
\name{insert_econ_datastore}
\alias{insert_econ_datastore}
\title{Inserts economy datastore data}
\usage{
insert_econ_datastore(log = "")
}
\arguments{
\item{log}{log file to save output to - defaults to}
}
\value{
Logical - TRUE for worked ok.
}
\description{
Inserts economy datastore data
}
\examples{
run_updates()

}
