% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_econ_labmarket.R
\name{insert_econ_labmarket}
\alias{insert_econ_labmarket}
\title{Inserts economy labour market file data}
\usage{
insert_econ_labmarket(log = "")
}
\arguments{
\item{log}{log file to save output to - defaults to}
}
\value{
Logical - TRUE for worked ok.
}
\description{
Inserts economy labour market file data
}
\examples{
run_updates()

}
