% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_econ_other.R
\name{insert_econ_other}
\alias{insert_econ_other}
\title{Inserts other economy data}
\usage{
insert_econ_other(log = "")
}
\arguments{
\item{log}{log file to save output to - defaults to output to console}
}
\value{
Logical - TRUE for worked ok.
}
\description{
Inserts other economy data
}
\examples{
run_updates()




}
