% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_inequalities_datastore.R
\name{insert_inequalities_datastore}
\alias{insert_inequalities_datastore}
\title{Inserts inequalities datastore}
\usage{
insert_inequalities_datastore(log = "")
}
\arguments{
\item{log}{log file to save output to - defaults to output to console}
}
\value{
Logical - TRUE for worked ok.
}
\description{
Inserts inequalities datastore
}
\examples{
run_updates()

}
