% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_subregional.R
\name{insert_subregional}
\alias{insert_subregional}
\title{Inserts data from surveys file maintained by Barry Fong}
\usage{
insert_subregional(log = "")
}
\arguments{
\item{log}{log file to save output to - defaults to}
}
\value{
Logical - TRUE for worked ok.
}
\description{
Inserts data from surveys file maintained by Barry Fong
}
\examples{
run_updates()




}
