% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_support_young_am.R
\name{insert_support_young_am}
\alias{insert_support_young_am}
\title{Inserts data from file created by Alice Major on young people}
\usage{
insert_support_young_am(log = "")
}
\arguments{
\item{log}{log file to save output to - defaults to}
}
\value{
Logical - TRUE for worked ok.
}
\description{
Inserts data from file created by Alice Major on young people
}
\examples{
run_updates()

}
