% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_support_young_other.R
\name{insert_support_young_other}
\alias{insert_support_young_other}
\title{Inserts data for young people from Datatstore}
\usage{
insert_support_young_other(log = "")
}
\arguments{
\item{log}{log file to save output to - defaults to}
}
\value{
Logical - TRUE for worked ok.
}
\description{
Inserts data for young people from Datatstore
}
\examples{
run_updates()

}
