% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbd_connect.R
\name{rdb_connect}
\alias{rdb_connect}
\title{Get database connection}
\usage{
rdb_connect()
}
\value{
#' A DBI dataconnect object
}
\description{
Get database connection
}
\examples{
conn <- rdb_connect()
}
