% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refresh_metadata.R
\name{refresh_metadata}
\alias{refresh_metadata}
\title{Refresh the metadata}
\usage{
refresh_metadata()
}
\value{
Logical - TRUE for worked ok
}
\description{
Refresh the metadata
}
\examples{
# my_median(1:12)
refresh_metadata()
}
